[#1065] Foremost·EventList 순수 뷰와 ViewModel 을 WidgetScenes 로 내린다
- `InlineSizeForemostEventView`·`SystemSizeForemostEventView`·`EventListView` 와 그 ViewModel 이 `WidgetScenes` 로 갔다. `EventTimeText.singleLineAttrText` 와 `ViewAppearance(google:_:)` 도 순수 뷰가 쓰는 것이라 함께 내렸다
- 할일 체크박스는 순수 뷰가 주입받는다 (FRAGO-2) — `TodoToggleIntent` 가 확장 전용 DI 를 써서 프레임워크로 못 내려간다. 그림을 그리는 `TodoToggleStyle`·`ForemostTodoToggleStyle` 만 내리고, `Toggle(intent:)` 로 감싼 버튼은 확장이 넘긴다
- `InlineSizeForemostEventView` 의 `.widgetURL`·`.widgetAccentable()` 을 엔트리 뷰로 올렸다. `DDayWidget` 이 `AccessoryWidgetBackground` 를 그렇게 다루는 것과 같은 형태다
- `SectionModel.Builder` 와 `EventListWidgetSize.init(_ family:)` 는 확장에 남겼다. 전자는 확장 소유 `CalendarEvents` 를, 후자는 `WidgetFamily` 를 받는다
