+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/ComposedWidget/EventAndMonthWidget/EventAndMonthWidget.swift
@@ -29,8 +29,11 @@ struct EventAndMonthWidgetView: View {
     var body: some View {
         switch self.entry.result {
         case .success(let model):
+            let colorSet = model.event.widgetSetting.background.colorSet(colorScheme == .light)
             HStack(spacing: 12) {
-                EventListView(model: model.event)
+                EventListView(model: model.event) { todo in
+                    AnyView(TodoToggleButton(todo: todo, colorSet: colorSet))
+                }
                 SingleMonthView(model: model.month)
             }
         case .failure(let error):
