+++ TodoCalendarApp/AppExtensions/Widget/Sources/Widgets/ForemostWidget/ForemostEventWidgetTimelineProvider.swift
@@ -13,6 +13,7 @@ import Optics
 import Domain
 import Extensions
 import CalendarPresentation
+import WidgetScenes
 
 
 struct ForemostEventWidgetTimelineProvider: TimelineProvider {
